changed strand, frame, and q_id to uint8_t instead of int; exchanged some resisting unsigned longs for size_t; renamed 63 searches in 64 searches, as they are actually doing 64 bit searches, except the Smith-Waterman, which is still 63 bit

+++ src/algo/16/16_simd_nw.c
@@ -203,9 +203,9 @@ static void aligncolumns_rest( __mxxxi * Sm, __mxxxi * hep, __mxxxi ** qp, __mxx
 }
 
 #ifdef __AVX2__
-void search_16_avx2_nw( p_s16info s, p_db_chunk chunk, p_minheap heap, p_node * overflow_list, int q_id ) {
+void search_16_avx2_nw( p_s16info s, p_db_chunk chunk, p_minheap heap, p_node * overflow_list, uint8_t q_id ) {
 #else
-void search_16_sse2_nw( p_s16info s, p_db_chunk chunk, p_minheap heap, p_node * overflow_list, int q_id ) {
+void search_16_sse2_nw( p_s16info s, p_db_chunk chunk, p_minheap heap, p_node * overflow_list, uint8_t q_id ) {
 #endif
 
 #ifdef DBG_COLLECT_MATRIX
